backend/requirements.txt
absl-py
fastapi
psycopg2-binary
python-dotenv
uvicorn
uvicorn[standard]
